Fix exchange format

+++ app/src/main/java/fr/iut_amiens/weatherapplication/openweathermap/WeatherResponse.java
@@ -472,16 +472,16 @@ public class WeatherResponse {
          * Rain volume for the last 3 hours
          */
         @SerializedName("3h")
-        private int threeHours;
+        private double threeHours;
 
         public Rain() {
         }
 
-        public int getThreeHours() {
+        public double getThreeHours() {
             return threeHours;
         }
 
-        public void setThreeHours(int threeHours) {
+        public void setThreeHours(double threeHours) {
             this.threeHours = threeHours;
         }
 
@@ -499,16 +499,16 @@ public class WeatherResponse {
          * Snow volume for the last 3 hours
          */
         @SerializedName("3h")
-        private int threeHours;
+        private double threeHours;
 
         public Snow() {
         }
 
-        public int getThreeHours() {
+        public double getThreeHours() {
             return threeHours;
         }
 
-        public void setThreeHours(int threeHours) {
+        public void setThreeHours(double threeHours) {
             this.threeHours = threeHours;
         }
